Moved _generate_unused_username to __init.py__
