+++ djangae/contrib/googleauth/backends/iap.py
@@ -6,32 +6,10 @@ from django.contrib.auth import get_user_model
 
 from djangae.contrib.googleauth.models import UserManager
 
-from . import _find_atomic_decorator
+from . import _find_atomic_decorator, _generate_unused_username
 from .base import BaseBackend
 
 User = get_user_model()
-
-
-def _generate_unused_username(ideal):
-    """
-        Check the database for a user with the specified username
-        and return either that ideal username, or an unused generated
-        one using the ideal username as a base
-    """
-
-    if not User.objects.filter(username_lower=ideal.lower()).exists():
-        return ideal
-
-    exists = True
-
-    # We use random digits rather than anything sequential to avoid any kind of
-    # attack vector to get this loop stuck
-    while exists:
-        random_digits = "".join([random.choice(string.digits) for x in range(5)])
-        username = "%s-%s" % (ideal, random_digits)
-        exists = User.objects.filter(username_lower=username.lower).exists()
-
-    return username
 
 
 class IAPBackend(BaseBackend):
